% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_project.R
\name{numeric_summary}
\alias{numeric_summary}
\title{This function creates a summary for a numeric vector}
\usage{
numeric_summary(x, na.rm = FALSE)
}
\arguments{
\item{creates}{the character of the given value}
}
\value{
the value
}
\description{
This function creates a summary for a numeric vector
}
